Modified dashboard UI. Refactored TopBarTitle to be suitable for custom organization
