+++ app/src/main/res/values/strings.xml
@@ -174,4 +174,5 @@ Privacy Policy
     <string name="tracking_progress_header">Tracking in progress</string>
     <string name="tracking_progress_message">Do not move the phone until tracking is completed</string>
     <string name="privacy_policy">Privacy Policy</string>
+    <string name="org">ORG.</string>
 </resources>
